decoder: Move Mapping interfaces to "interfaces.ts"
