+++ src/decoder.ts
@@ -1,21 +1,5 @@
 import { Delegate } from './mappings-decoder';
-
-export interface Mapping {
-  fieldCount: number;
-  col: number;
-  src: number;
-  srcLine: number;
-  srcCol: number;
-  name: number;
-}
-
-export interface LineMappings {
-  mappings: Array<Mapping>;
-}
-
-export interface FileMappings {
-  lines: Array<LineMappings>;
-}
+import { Mapping, FileMappings } from './interfaces';
 
 export default class Decoder implements Delegate {
   currentLine: Array<Mapping> = [];
